+++ sdk/python/capsem/__init__.py
@@ -3,4 +3,5 @@
 from ._transport import HttpError as HttpError
 from .execution import decode_exec_output as decode_exec_output
 from .hypervisor import Hypervisor as Hypervisor
+from .options import ContainerOptions as ContainerOptions
 from .vm import VM as VM